refactor: atom, selector 관심사에 따라 재분류 및 폴더링
