+++ src/components/CartItem/CartItem.tsx
@@ -2,7 +2,7 @@ import { useRecoilState } from 'recoil';
 import CheckButton from '../Button/CheckButton/CheckButton';
 import QuantityContainer from '../Container/QuantityContainer/QuantityContainer';
 import type { TCartItem } from '../../types/CartItem.type';
-import { selectedCartItemListSelector } from '../../recoil/selectors/selectors';
+import { selectedCartItemListSelector } from '../../pages/ShoppingCartPage/recoil/selector/selectedCartItemListSelector';
 
 import * as S from './CartItem.style';
