+++ src/pages/OrderConfirmPage/OrderConfirmPage.tsx
@@ -3,11 +3,11 @@ import { Navigate } from 'react-router-dom';
 import Header from '../../components/Header/Header';
 import TitleContainer from '../../components/Container/TitleContainer/TitleContainer';
 import SubmitButton from '../../components/Button/SubmitButton/SubmitButton';
-import { selectedCartItemListState } from '../../recoil/atoms/atoms';
+import { selectedCartItemListState } from '../ShoppingCartPage/recoil/atom/';
 import {
   selectedCartItemListTotalPriceSelector,
   selectedCartItemListTotalCountSelector,
-} from '../../recoil/selectors/selectors';
+} from '../ShoppingCartPage/recoil/selector/selectedCartItemListSelector';
 import { calculateDeliveryFee } from '../../utils/calculateDeliveryFee';
 import { PATHS } from '../../constants/PATHS';
 import * as S from './OrderConfirmPage.style';
