+++ src/pages/ShoppingCartPage/recoil/atom/selectedCartItemListState.ts
@@ -1,5 +1,5 @@
 import { atom, AtomEffect } from 'recoil';
-import type { TCartItem } from '../../types/CartItem.type';
+import type { TCartItem } from '../../../../types/CartItem.type';
 
 const localStorageEffect: <T>(key: string) => AtomEffect<T> =
   (key: string) =>
