+++ src/pages/ShoppingCartPage/recoil/selector/selectors.test.ts
@@ -1,12 +1,10 @@
 import { renderHook, act } from '@testing-library/react';
 import { RecoilRoot, useSetRecoilState, useRecoilState, useRecoilValue } from 'recoil';
-import { selectedCartItemListState } from '../atoms/atoms';
-import {
-  selectedCartItemListSelector,
-  selectedCartItemListTotalPriceSelector,
-  selectedCartItemListTotalCountSelector,
-} from './selectors';
-import { cartItemListMockData, newYujoCartItemData, newKyleCartItemData } from '../../data/cartItemListMockData';
+import { selectedCartItemListState } from '../atom/selectedCartItemListState';
+import { selectedCartItemListSelector } from './selectedCartItemListSelector';
+import { selectedCartItemListTotalPriceSelector } from './selectedCartItemListTotalPriceSelector';
+import { selectedCartItemListTotalCountSelector } from './selectedCartItemListTotalCountSelector';
+import { cartItemListMockData, newYujoCartItemData, newKyleCartItemData } from '../../../../data/cartItemListMockData';
 
 describe('selectedCartItemListSelector', () => {
   it('개별 상품을 선택할 시 선택된 상품이 배열에 삽입된다.', () => {
